Refactor GuiSystem by extracting tab content into separate files
Break up the large GuiSystem.cpp file (~2570 lines) by extracting each
tab's render content into its own compilation unit:

- GuiTimeTab: Time of day, date, location, moon phase, eclipse controls
- GuiWeatherTab: Weather type, snow, wind settings
- GuiEnvironmentTab: Fog, atmosphere, clouds, grass settings
- GuiPostFXTab: HDR, bloom, god rays, volumetric fog quality
- GuiTerrainTab: LOD parameters, meshlets, optimizations, streaming
- GuiWaterTab: Water level, waves, appearance, presets
- GuiDebugTab: Debug visualizations, occlusion culling, shortcuts
- GuiProfilerTab: GPU/CPU timing breakdown, frame budget
- GuiIKTab: IK controls, skeleton overlay rendering
- GuiPlayerTab: Cape settings

GuiSystem.cpp is now reduced to core initialization, styling, and
orchestration code (~575 lines). Tab state that needs persistence
is passed through structs (EnvironmentTabState, IKDebugSettings,
PlayerSettings).
