+++ src/GuiDebugTab.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class Renderer;
+
+namespace GuiDebugTab {
+    void render(Renderer& renderer);
+}
