+++ src/GuiPlayerTab.cpp
@@ -1,0 +1,42 @@
+#include "GuiPlayerTab.h"
+#include "Renderer.h"
+
+#include <imgui.h>
+
+void GuiPlayerTab::render(Renderer& renderer, PlayerSettings& settings) {
+    ImGui::Spacing();
+
+    auto& sceneBuilder = renderer.getSceneBuilder();
+    if (!sceneBuilder.hasCharacter()) {
+        ImGui::TextDisabled("No animated character loaded");
+        return;
+    }
+
+    // Cape section
+    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.9f, 0.5f, 0.5f, 1.0f));
+    ImGui::Text("CAPE");
+    ImGui::PopStyleColor();
+
+    ImGui::Checkbox("Enable Cape", &settings.capeEnabled);
+    if (ImGui::IsItemHovered()) {
+        ImGui::SetTooltip("Toggle cape visibility and simulation");
+    }
+
+    ImGui::Checkbox("Show Cape Colliders", &settings.showCapeColliders);
+    if (ImGui::IsItemHovered()) {
+        ImGui::SetTooltip("Visualize body colliders used for cape collision");
+    }
+
+    ImGui::Spacing();
+    ImGui::Separator();
+    ImGui::Spacing();
+
+    // Cape info
+    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.7f, 0.7f, 1.0f));
+    ImGui::Text("CAPE INFO");
+    ImGui::PopStyleColor();
+
+    ImGui::BulletText("Cloth simulation: Verlet integration");
+    ImGui::BulletText("Body colliders: Spheres + Capsules");
+    ImGui::BulletText("Attachments: Shoulders + Upper back");
+}
