+++ src/GuiPostFXTab.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class Renderer;
+
+namespace GuiPostFXTab {
+    void render(Renderer& renderer);
+}
