+++ src/GuiProfilerTab.cpp
@@ -1,0 +1,161 @@
+#include "GuiProfilerTab.h"
+#include "Renderer.h"
+
+#include <imgui.h>
+#include <algorithm>
+#include <cstdio>
+
+void GuiProfilerTab::render(Renderer& renderer) {
+    ImGui::Spacing();
+
+    auto& profiler = renderer.getProfiler();
+
+    // Enable/disable toggle
+    bool enabled = profiler.isEnabled();
+    if (ImGui::Checkbox("Enable Profiling", &enabled)) {
+        profiler.setEnabled(enabled);
+    }
+
+    if (!enabled) {
+        ImGui::TextDisabled("Profiling disabled");
+        return;
+    }
+
+    ImGui::Spacing();
+    ImGui::Separator();
+    ImGui::Spacing();
+
+    // GPU Profiling Section
+    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.4f, 0.8f, 1.0f, 1.0f));
+    ImGui::Text("GPU TIMING");
+    ImGui::PopStyleColor();
+
+    const auto& gpuStats = profiler.getSmoothedGpuResults();
+
+    if (gpuStats.zones.empty()) {
+        ImGui::TextDisabled("No GPU data yet (waiting for frames)");
+    } else {
+        // Total GPU time
+        ImGui::Text("Total GPU: %.2f ms", gpuStats.totalGpuTimeMs);
+
+        ImGui::Spacing();
+
+        // GPU timing breakdown table
+        if (ImGui::BeginTable("GPUTimings", 3, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
+            ImGui::TableSetupColumn("Pass", ImGuiTableColumnFlags_WidthStretch);
+            ImGui::TableSetupColumn("Time (ms)", ImGuiTableColumnFlags_WidthFixed, 70.0f);
+            ImGui::TableSetupColumn("%", ImGuiTableColumnFlags_WidthFixed, 50.0f);
+            ImGui::TableHeadersRow();
+
+            for (const auto& zone : gpuStats.zones) {
+                ImGui::TableNextRow();
+
+                ImGui::TableNextColumn();
+                ImGui::Text("%s", zone.name.c_str());
+
+                ImGui::TableNextColumn();
+                ImGui::Text("%.2f", zone.gpuTimeMs);
+
+                ImGui::TableNextColumn();
+                // Color code by percentage
+                if (zone.percentOfFrame > 30.0f) {
+                    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.4f, 0.4f, 1.0f));
+                } else if (zone.percentOfFrame > 15.0f) {
+                    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.8f, 0.4f, 1.0f));
+                } else {
+                    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.4f, 1.0f, 0.4f, 1.0f));
+                }
+                ImGui::Text("%.1f%%", zone.percentOfFrame);
+                ImGui::PopStyleColor();
+            }
+
+            ImGui::EndTable();
+        }
+
+        // Visual bar chart of GPU zones
+        ImGui::Spacing();
+        float maxTime = gpuStats.totalGpuTimeMs;
+        for (const auto& zone : gpuStats.zones) {
+            float fraction = (maxTime > 0.0f) ? (zone.gpuTimeMs / maxTime) : 0.0f;
+            ImGui::ProgressBar(fraction, ImVec2(-1, 0), zone.name.c_str());
+        }
+    }
+
+    ImGui::Spacing();
+    ImGui::Separator();
+    ImGui::Spacing();
+
+    // CPU Profiling Section
+    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.8f, 0.4f, 1.0f));
+    ImGui::Text("CPU TIMING");
+    ImGui::PopStyleColor();
+
+    const auto& cpuStats = profiler.getSmoothedCpuResults();
+
+    if (cpuStats.zones.empty()) {
+        ImGui::TextDisabled("No CPU data yet");
+    } else {
+        // Total CPU time
+        ImGui::Text("Total CPU: %.2f ms", cpuStats.totalCpuTimeMs);
+
+        ImGui::Spacing();
+
+        // CPU timing breakdown table
+        if (ImGui::BeginTable("CPUTimings", 3, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
+            ImGui::TableSetupColumn("Zone", ImGuiTableColumnFlags_WidthStretch);
+            ImGui::TableSetupColumn("Time (ms)", ImGuiTableColumnFlags_WidthFixed, 70.0f);
+            ImGui::TableSetupColumn("%", ImGuiTableColumnFlags_WidthFixed, 50.0f);
+            ImGui::TableHeadersRow();
+
+            for (const auto& zone : cpuStats.zones) {
+                ImGui::TableNextRow();
+
+                ImGui::TableNextColumn();
+                ImGui::Text("%s", zone.name.c_str());
+
+                ImGui::TableNextColumn();
+                ImGui::Text("%.3f", zone.cpuTimeMs);
+
+                ImGui::TableNextColumn();
+                ImGui::Text("%.1f%%", zone.percentOfFrame);
+            }
+
+            ImGui::EndTable();
+        }
+    }
+
+    ImGui::Spacing();
+    ImGui::Separator();
+    ImGui::Spacing();
+
+    // Frame budget indicator
+    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.8f, 0.8f, 1.0f, 1.0f));
+    ImGui::Text("FRAME BUDGET");
+    ImGui::PopStyleColor();
+
+    float targetMs = 16.67f;  // 60 FPS target
+    float gpuTime = gpuStats.totalGpuTimeMs;
+    float cpuTime = cpuStats.totalCpuTimeMs;
+    float maxTime = std::max(gpuTime, cpuTime);
+
+    // Budget bar
+    float budgetUsed = maxTime / targetMs;
+    ImVec4 budgetColor;
+    if (budgetUsed < 0.8f) {
+        budgetColor = ImVec4(0.4f, 1.0f, 0.4f, 1.0f);  // Green
+    } else if (budgetUsed < 1.0f) {
+        budgetColor = ImVec4(1.0f, 0.8f, 0.4f, 1.0f);  // Yellow
+    } else {
+        budgetColor = ImVec4(1.0f, 0.4f, 0.4f, 1.0f);  // Red
+    }
+
+    ImGui::PushStyleColor(ImGuiCol_PlotHistogram, budgetColor);
+    char budgetText[64];
+    snprintf(budgetText, sizeof(budgetText), "%.1f / %.1f ms (%.0f%%)",
+             maxTime, targetMs, budgetUsed * 100.0f);
+    ImGui::ProgressBar(std::min(budgetUsed, 1.5f) / 1.5f, ImVec2(-1, 20), budgetText);
+    ImGui::PopStyleColor();
+
+    ImGui::Text("GPU Bound: %s", (gpuTime > cpuTime) ? "Yes" : "No");
+    ImGui::Text("CPU Bound: %s", (cpuTime > gpuTime) ? "Yes" : "No");
+}
