+++ src/GuiProfilerTab.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class Renderer;
+
+namespace GuiProfilerTab {
+    void render(Renderer& renderer);
+}
