+++ src/GuiSystem.cpp
@@ -1,13 +1,17 @@
 #include "GuiSystem.h"
 #include "Renderer.h"
 #include "Camera.h"
-#include "AtmosphereLUTSystem.h"
-#include "AnimatedCharacter.h"
-#include "PlayerCape.h"
-#include "DebugLineSystem.h"
-#ifdef JPH_DEBUG_RENDERER
-#include "PhysicsDebugRenderer.h"
-#endif
+
+#include "GuiTimeTab.h"
+#include "GuiWeatherTab.h"
+#include "GuiEnvironmentTab.h"
+#include "GuiPostFXTab.h"
+#include "GuiTerrainTab.h"
+#include "GuiWaterTab.h"
+#include "GuiDebugTab.h"
+#include "GuiProfilerTab.h"
+#include "GuiIKTab.h"
+#include "GuiPlayerTab.h"
 
 #include <imgui.h>
 #include <imgui_impl_sdl3.h>
@@ -246,43 +250,43 @@ void GuiSystem::render(Renderer& renderer, const Camera& camera, float deltaTime
 
         if (ImGui::BeginTabBar("ControlTabs")) {
             if (ImGui::BeginTabItem("Time")) {
-                renderTimeSection(renderer);
+                GuiTimeTab::render(renderer);
                 ImGui::EndTabItem();
             }
             if (ImGui::BeginTabItem("Weather")) {
-                renderWeatherSection(renderer);
+                GuiWeatherTab::render(renderer);
                 ImGui::EndTabItem();
             }
             if (ImGui::BeginTabItem("Environment")) {
-                renderEnvironmentSection(renderer);
+                GuiEnvironmentTab::render(renderer, environmentTabState);
                 ImGui::EndTabItem();
             }
             if (ImGui::BeginTabItem("Post FX")) {
-                renderPostProcessSection(renderer);
+                GuiPostFXTab::render(renderer);
                 ImGui::EndTabItem();
             }
             if (ImGui::BeginTabItem("Terrain")) {
-                renderTerrainSection(renderer);
+                GuiTerrainTab::render(renderer);
                 ImGui::EndTabItem();
             }
             if (ImGui::BeginTabItem("Water")) {
-                renderWaterSection(renderer);
+                GuiWaterTab::render(renderer);
                 ImGui::EndTabItem();
             }
             if (ImGui::BeginTabItem("Player")) {
-                renderPlayerSection(renderer);
+                GuiPlayerTab::render(renderer, playerSettings);
                 ImGui::EndTabItem();
             }
             if (ImGui::BeginTabItem("IK")) {
-                renderIKSection(renderer, camera);
+                GuiIKTab::render(renderer, camera, ikDebugSettings);
                 ImGui::EndTabItem();
             }
             if (ImGui::BeginTabItem("Debug")) {
-                renderDebugSection(renderer);
+                GuiDebugTab::render(renderer);
                 ImGui::EndTabItem();
             }
             if (ImGui::BeginTabItem("Profiler")) {
-                renderProfilerSection(renderer);
+                GuiProfilerTab::render(renderer);
                 ImGui::EndTabItem();
             }
             ImGui::EndTabBar();
@@ -300,7 +304,7 @@ void GuiSystem::render(Renderer& renderer, const Camera& camera, float deltaTime
 
     // Skeleton/IK debug overlay
     if (ikDebugSettings.showSkeleton || ikDebugSettings.showIKTargets) {
-        renderSkeletonOverlay(renderer, camera);
+        GuiIKTab::renderSkeletonOverlay(renderer, camera, ikDebugSettings, playerSettings.showCapeColliders);
     }
 
     // Tree editor as separate window
@@ -397,1574 +401,6 @@ void GuiSystem::renderDashboard(Renderer& renderer, const Camera& camera, float 
     if (ImGui::Button(treeEditorGui.isVisible() ? "Hide Tree Editor (F2)" : "Show Tree Editor (F2)", ImVec2(-1, 0))) {
         treeEditorGui.toggleVisibility();
     }
-}
-
-void GuiSystem::renderTimeSection(Renderer& renderer) {
-    ImGui::Spacing();
-
-    // Time of day slider
-    float timeOfDay = renderer.getTimeOfDay();
-    if (ImGui::SliderFloat("Time of Day", &timeOfDay, 0.0f, 1.0f, "%.3f")) {
-        renderer.setTimeOfDay(timeOfDay);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("0.0 = Midnight, 0.25 = Sunrise, 0.5 = Noon, 0.75 = Sunset");
-    }
-
-    // Quick time buttons
-    ImGui::Text("Presets:");
-    ImGui::SameLine();
-    if (ImGui::Button("Dawn")) renderer.setTimeOfDay(0.25f);
-    ImGui::SameLine();
-    if (ImGui::Button("Noon")) renderer.setTimeOfDay(0.5f);
-    ImGui::SameLine();
-    if (ImGui::Button("Dusk")) renderer.setTimeOfDay(0.75f);
-    ImGui::SameLine();
-    if (ImGui::Button("Night")) renderer.setTimeOfDay(0.0f);
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Time scale
-    float timeScale = renderer.getTimeScale();
-    if (ImGui::SliderFloat("Time Scale", &timeScale, 0.0f, 100.0f, "%.1fx", ImGuiSliderFlags_Logarithmic)) {
-        renderer.setTimeScale(timeScale);
-    }
-
-    if (ImGui::Button("Resume Real-Time")) {
-        renderer.resumeAutoTime();
-        renderer.setTimeScale(1.0f);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Date controls
-    ImGui::Text("Date (affects sun position):");
-    int year = renderer.getCurrentYear();
-    int month = renderer.getCurrentMonth();
-    int day = renderer.getCurrentDay();
-
-    bool dateChanged = false;
-    ImGui::SetNextItemWidth(80);
-    if (ImGui::InputInt("Year", &year, 1, 10)) dateChanged = true;
-    ImGui::SameLine();
-    ImGui::SetNextItemWidth(60);
-    if (ImGui::InputInt("Month", &month, 1, 1)) dateChanged = true;
-    ImGui::SameLine();
-    ImGui::SetNextItemWidth(60);
-    if (ImGui::InputInt("Day", &day, 1, 1)) dateChanged = true;
-
-    if (dateChanged) {
-        month = std::clamp(month, 1, 12);
-        day = std::clamp(day, 1, 31);
-        renderer.setDate(year, month, day);
-    }
-
-    // Season presets
-    ImGui::Text("Season:");
-    ImGui::SameLine();
-    if (ImGui::Button("Spring")) renderer.setDate(renderer.getCurrentYear(), 3, 20);
-    ImGui::SameLine();
-    if (ImGui::Button("Summer")) renderer.setDate(renderer.getCurrentYear(), 6, 21);
-    ImGui::SameLine();
-    if (ImGui::Button("Autumn")) renderer.setDate(renderer.getCurrentYear(), 9, 22);
-    ImGui::SameLine();
-    if (ImGui::Button("Winter")) renderer.setDate(renderer.getCurrentYear(), 12, 21);
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Location
-    GeographicLocation loc = renderer.getLocation();
-    float lat = static_cast<float>(loc.latitude);
-    float lon = static_cast<float>(loc.longitude);
-    bool locChanged = false;
-
-    if (ImGui::SliderFloat("Latitude", &lat, -90.0f, 90.0f, "%.1f")) locChanged = true;
-    if (ImGui::SliderFloat("Longitude", &lon, -180.0f, 180.0f, "%.1f")) locChanged = true;
-
-    if (locChanged) {
-        renderer.setLocation({static_cast<double>(lat), static_cast<double>(lon)});
-    }
-
-    // Location presets
-    ImGui::Text("Location:");
-    if (ImGui::Button("London")) {
-        renderer.setLocation({51.5f, -0.1f});
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("New York")) {
-        renderer.setLocation({40.7f, -74.0f});
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Tokyo")) {
-        renderer.setLocation({35.7f, 139.7f});
-    }
-    if (ImGui::Button("Sydney")) {
-        renderer.setLocation({-33.9f, 151.2f});
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Arctic")) {
-        renderer.setLocation({71.0f, 25.0f});
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Equator")) {
-        renderer.setLocation({0.0f, 0.0f});
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Moon Phase Controls
-    ImGui::Text("Moon Phase:");
-
-    // Display current moon phase
-    float currentPhase = renderer.getCurrentMoonPhase();
-    const char* phaseNames[] = { "New Moon", "Waxing Crescent", "First Quarter", "Waxing Gibbous",
-                                 "Full Moon", "Waning Gibbous", "Last Quarter", "Waning Crescent" };
-    int phaseIndex = static_cast<int>(currentPhase * 8.0f) % 8;
-    ImGui::Text("Current: %s (%.2f)", phaseNames[phaseIndex], currentPhase);
-
-    // Override checkbox
-    bool overrideEnabled = renderer.isMoonPhaseOverrideEnabled();
-    if (ImGui::Checkbox("Override Moon Phase", &overrideEnabled)) {
-        renderer.setMoonPhaseOverride(overrideEnabled);
-    }
-
-    // Manual phase slider (only active when override is enabled)
-    if (overrideEnabled) {
-        float manualPhase = renderer.getMoonPhase();
-        if (ImGui::SliderFloat("Moon Phase", &manualPhase, 0.0f, 1.0f, "%.3f")) {
-            renderer.setMoonPhase(manualPhase);
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0.0 = New Moon, 0.25 = First Quarter, 0.5 = Full Moon, 0.75 = Last Quarter");
-        }
-
-        // Quick phase buttons
-        ImGui::Text("Presets:");
-        ImGui::SameLine();
-        if (ImGui::Button("New")) renderer.setMoonPhase(0.0f);
-        ImGui::SameLine();
-        if (ImGui::Button("1st Q")) renderer.setMoonPhase(0.25f);
-        ImGui::SameLine();
-        if (ImGui::Button("Full")) renderer.setMoonPhase(0.5f);
-        ImGui::SameLine();
-        if (ImGui::Button("3rd Q")) renderer.setMoonPhase(0.75f);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Moon Brightness Controls
-    ImGui::Text("Moon Brightness:");
-
-    float moonBrightness = renderer.getMoonBrightness();
-    if (ImGui::SliderFloat("Light Intensity", &moonBrightness, 0.0f, 5.0f, "%.2f")) {
-        renderer.setMoonBrightness(moonBrightness);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Multiplier for moonlight intensity on terrain (0-5, default 1.0)");
-    }
-
-    float moonDiscIntensity = renderer.getMoonDiscIntensity();
-    if (ImGui::SliderFloat("Disc Intensity", &moonDiscIntensity, 0.0f, 50.0f, "%.1f")) {
-        renderer.setMoonDiscIntensity(moonDiscIntensity);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Visual brightness of moon disc in sky (0-50, default 20)");
-    }
-
-    float moonEarthshine = renderer.getMoonEarthshine();
-    if (ImGui::SliderFloat("Earthshine", &moonEarthshine, 0.0f, 0.2f, "%.3f")) {
-        renderer.setMoonEarthshine(moonEarthshine);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Visibility of dark side during crescent phases (0-0.2, default 0.02)");
-    }
-
-    // Brightness presets
-    ImGui::Text("Presets:");
-    ImGui::SameLine();
-    if (ImGui::Button("Dim")) {
-        renderer.setMoonBrightness(0.5f);
-        renderer.setMoonDiscIntensity(10.0f);
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Normal")) {
-        renderer.setMoonBrightness(1.0f);
-        renderer.setMoonDiscIntensity(20.0f);
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Bright")) {
-        renderer.setMoonBrightness(2.0f);
-        renderer.setMoonDiscIntensity(35.0f);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Eclipse Controls
-    ImGui::Text("Solar Eclipse:");
-
-    bool eclipseEnabled = renderer.isEclipseEnabled();
-    if (ImGui::Checkbox("Enable Eclipse", &eclipseEnabled)) {
-        renderer.setEclipseEnabled(eclipseEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Simulates a solar eclipse with the moon passing in front of the sun");
-    }
-
-    if (eclipseEnabled) {
-        float eclipseAmount = renderer.getEclipseAmount();
-        if (ImGui::SliderFloat("Eclipse Amount", &eclipseAmount, 0.0f, 1.0f, "%.3f")) {
-            renderer.setEclipseAmount(eclipseAmount);
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0.0 = No eclipse, 1.0 = Total eclipse");
-        }
-
-        // Eclipse presets
-        ImGui::Text("Presets:");
-        ImGui::SameLine();
-        if (ImGui::Button("Partial")) renderer.setEclipseAmount(0.5f);
-        ImGui::SameLine();
-        if (ImGui::Button("Annular")) renderer.setEclipseAmount(0.85f);
-        ImGui::SameLine();
-        if (ImGui::Button("Total")) renderer.setEclipseAmount(1.0f);
-    }
-}
-
-void GuiSystem::renderWeatherSection(Renderer& renderer) {
-    ImGui::Spacing();
-
-    // Weather type
-    const char* weatherTypes[] = { "Rain", "Snow" };
-    int weatherType = static_cast<int>(renderer.getWeatherType());
-    if (ImGui::Combo("Weather Type", &weatherType, weatherTypes, 2)) {
-        renderer.setWeatherType(static_cast<uint32_t>(weatherType));
-    }
-
-    // Intensity
-    float intensity = renderer.getIntensity();
-    if (ImGui::SliderFloat("Intensity", &intensity, 0.0f, 1.0f)) {
-        renderer.setWeatherIntensity(intensity);
-    }
-
-    // Quick intensity buttons
-    ImGui::Text("Presets:");
-    ImGui::SameLine();
-    if (ImGui::Button("Clear")) renderer.setWeatherIntensity(0.0f);
-    ImGui::SameLine();
-    if (ImGui::Button("Light")) renderer.setWeatherIntensity(0.3f);
-    ImGui::SameLine();
-    if (ImGui::Button("Medium")) renderer.setWeatherIntensity(0.6f);
-    ImGui::SameLine();
-    if (ImGui::Button("Heavy")) renderer.setWeatherIntensity(1.0f);
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Snow coverage
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.85f, 1.0f, 1.0f));
-    ImGui::Text("SNOW COVERAGE");
-    ImGui::PopStyleColor();
-
-    float snowAmount = renderer.getSnowAmount();
-    if (ImGui::SliderFloat("Snow Amount", &snowAmount, 0.0f, 1.0f)) {
-        renderer.setSnowAmount(snowAmount);
-    }
-
-    glm::vec3 snowColor = renderer.getSnowColor();
-    float sc[3] = {snowColor.r, snowColor.g, snowColor.b};
-    if (ImGui::ColorEdit3("Snow Color", sc)) {
-        renderer.setSnowColor(glm::vec3(sc[0], sc[1], sc[2]));
-    }
-
-    // Environment settings for snow
-    auto& env = renderer.getEnvironmentSettings();
-
-    if (ImGui::SliderFloat("Snow Roughness", &env.snowRoughness, 0.0f, 1.0f)) {}
-    if (ImGui::SliderFloat("Accumulation Rate", &env.snowAccumulationRate, 0.0f, 1.0f)) {}
-    if (ImGui::SliderFloat("Melt Rate", &env.snowMeltRate, 0.0f, 1.0f)) {}
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Wind settings
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.6f, 0.9f, 0.7f, 1.0f));
-    ImGui::Text("WIND");
-    ImGui::PopStyleColor();
-
-    float windDir[2] = {env.windDirection.x, env.windDirection.y};
-    if (ImGui::SliderFloat2("Direction", windDir, -1.0f, 1.0f)) {
-        env.windDirection = glm::vec2(windDir[0], windDir[1]);
-    }
-
-    if (ImGui::SliderFloat("Strength", &env.windStrength, 0.0f, 3.0f)) {}
-    if (ImGui::SliderFloat("Speed", &env.windSpeed, 0.0f, 5.0f)) {}
-    if (ImGui::SliderFloat("Gust Frequency", &env.gustFrequency, 0.0f, 2.0f)) {}
-    if (ImGui::SliderFloat("Gust Amplitude", &env.gustAmplitude, 0.0f, 2.0f)) {}
-}
-
-void GuiSystem::renderEnvironmentSection(Renderer& renderer) {
-    ImGui::Spacing();
-
-    // ========== FROXEL VOLUMETRIC FOG ==========
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.7f, 0.9f, 1.0f));
-    ImGui::Text("FROXEL VOLUMETRIC FOG");
-    ImGui::PopStyleColor();
-
-    bool fogEnabled = renderer.isFogEnabled();
-    if (ImGui::Checkbox("Enable Froxel Fog", &fogEnabled)) {
-        renderer.setFogEnabled(fogEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Frustum-aligned voxel grid volumetric fog with temporal reprojection");
-    }
-
-    if (fogEnabled) {
-        // Main fog parameters - wide ranges for extreme testing
-        float fogDensity = renderer.getFogDensity();
-        if (ImGui::SliderFloat("Fog Density", &fogDensity, 0.0f, 1.0f, "%.4f", ImGuiSliderFlags_Logarithmic)) {
-            renderer.setFogDensity(fogDensity);
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0 = no fog, 1 = extremely dense (logarithmic scale)");
-        }
-
-        float fogAbsorption = renderer.getFogAbsorption();
-        if (ImGui::SliderFloat("Absorption", &fogAbsorption, 0.0f, 1.0f, "%.4f", ImGuiSliderFlags_Logarithmic)) {
-            renderer.setFogAbsorption(fogAbsorption);
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("Light absorption coefficient (0 = transparent, 1 = opaque fog)");
-        }
-
-        float fogBaseHeight = renderer.getFogBaseHeight();
-        if (ImGui::SliderFloat("Base Height", &fogBaseHeight, -500.0f, 500.0f, "%.1f")) {
-            renderer.setFogBaseHeight(fogBaseHeight);
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("Height where fog density is maximum");
-        }
-
-        float fogScaleHeight = renderer.getFogScaleHeight();
-        if (ImGui::SliderFloat("Scale Height", &fogScaleHeight, 0.1f, 2000.0f, "%.1f", ImGuiSliderFlags_Logarithmic)) {
-            renderer.setFogScaleHeight(fogScaleHeight);
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("Exponential falloff (0.1 = thin layer, 2000 = fog everywhere)");
-        }
-
-        float volumetricFar = renderer.getVolumetricFarPlane();
-        if (ImGui::SliderFloat("Far Plane", &volumetricFar, 10.0f, 5000.0f, "%.0f", ImGuiSliderFlags_Logarithmic)) {
-            renderer.setVolumetricFarPlane(volumetricFar);
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("Volumetric range (10 = close only, 5000 = entire scene)");
-        }
-
-        float temporalBlend = renderer.getTemporalBlend();
-        if (ImGui::SliderFloat("Temporal Blend", &temporalBlend, 0.0f, 0.999f, "%.3f")) {
-            renderer.setTemporalBlend(temporalBlend);
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0 = no temporal filtering (noisy), 0.999 = extreme smoothing (ghosting)");
-        }
-
-        // Quick presets for common scenarios
-        ImGui::Text("Presets:");
-        ImGui::SameLine();
-        if (ImGui::Button("Clear##froxel")) {
-            renderer.setFogDensity(0.0f);
-            renderer.setLayerDensity(0.0f);
-        }
-        ImGui::SameLine();
-        if (ImGui::Button("Light##froxel")) {
-            renderer.setFogDensity(0.005f);
-            renderer.setFogAbsorption(0.005f);
-            renderer.setFogScaleHeight(100.0f);
-        }
-        ImGui::SameLine();
-        if (ImGui::Button("Dense##froxel")) {
-            renderer.setFogDensity(0.03f);
-            renderer.setFogAbsorption(0.02f);
-            renderer.setFogScaleHeight(50.0f);
-        }
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // ========== HEIGHT FOG LAYER ==========
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.6f, 0.8f, 0.9f, 1.0f));
-    ImGui::Text("HEIGHT FOG LAYER");
-    ImGui::PopStyleColor();
-
-    if (fogEnabled) {
-        // Enable toggle for height fog layer
-        if (ImGui::Checkbox("Enable Height Fog", &heightFogEnabled)) {
-            if (heightFogEnabled) {
-                // Restore cached density
-                renderer.setLayerDensity(cachedLayerDensity);
-            } else {
-                // Cache current density and zero it out
-                cachedLayerDensity = renderer.getLayerDensity();
-                if (cachedLayerDensity < 0.001f) cachedLayerDensity = 0.02f;  // Ensure valid restore value
-                renderer.setLayerDensity(0.0f);
-            }
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("Toggle ground-hugging fog layer");
-        }
-
-        if (heightFogEnabled) {
-            float layerHeight = renderer.getLayerHeight();
-            if (ImGui::SliderFloat("Layer Height", &layerHeight, -200.0f, 500.0f, "%.1f")) {
-                renderer.setLayerHeight(layerHeight);
-            }
-            if (ImGui::IsItemHovered()) {
-                ImGui::SetTooltip("Top of ground fog layer (-200 = below ground, 500 = high altitude cloud)");
-            }
-
-            float layerThickness = renderer.getLayerThickness();
-            if (ImGui::SliderFloat("Layer Thickness", &layerThickness, 0.1f, 500.0f, "%.1f", ImGuiSliderFlags_Logarithmic)) {
-                renderer.setLayerThickness(layerThickness);
-            }
-            if (ImGui::IsItemHovered()) {
-                ImGui::SetTooltip("Vertical extent (0.1 = paper thin, 500 = massive fog bank)");
-            }
-
-            float layerDensity = renderer.getLayerDensity();
-            if (ImGui::SliderFloat("Layer Density", &layerDensity, 0.0f, 1.0f, "%.4f", ImGuiSliderFlags_Logarithmic)) {
-                renderer.setLayerDensity(layerDensity);
-                cachedLayerDensity = layerDensity;  // Update cache when manually changed
-            }
-            if (ImGui::IsItemHovered()) {
-                ImGui::SetTooltip("0 = invisible, 1 = completely opaque (logarithmic)");
-            }
-
-            // Quick presets
-            ImGui::Text("Presets:");
-            ImGui::SameLine();
-            if (ImGui::Button("Valley##layer")) {
-                renderer.setLayerHeight(20.0f);
-                renderer.setLayerThickness(30.0f);
-                renderer.setLayerDensity(0.03f);
-                cachedLayerDensity = 0.03f;
-            }
-            ImGui::SameLine();
-            if (ImGui::Button("Thick Mist##layer")) {
-                renderer.setLayerHeight(10.0f);
-                renderer.setLayerThickness(15.0f);
-                renderer.setLayerDensity(0.1f);
-                cachedLayerDensity = 0.1f;
-            }
-        }
-    } else {
-        ImGui::TextDisabled("Enable Froxel Fog to access height fog settings");
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // ========== ATMOSPHERIC SCATTERING ==========
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.5f, 0.7f, 1.0f, 1.0f));
-    ImGui::Text("ATMOSPHERIC SCATTERING");
-    ImGui::PopStyleColor();
-
-    AtmosphereParams atmosParams = renderer.getAtmosphereParams();
-    bool atmosChanged = false;
-
-    // Enable toggle for atmospheric scattering
-    if (ImGui::Checkbox("Enable Atmosphere", &atmosphereEnabled)) {
-        if (atmosphereEnabled) {
-            // Restore cached values
-            atmosParams.rayleighScatteringBase = glm::vec3(5.802e-3f, 13.558e-3f, 33.1e-3f) * (cachedRayleighScale / 13.558f);
-            atmosParams.mieScatteringBase = cachedMieScale / 1000.0f;
-            atmosChanged = true;
-        } else {
-            // Cache current values and zero out scattering
-            cachedRayleighScale = atmosParams.rayleighScatteringBase.y * 1000.0f;
-            cachedMieScale = atmosParams.mieScatteringBase * 1000.0f;
-            if (cachedRayleighScale < 0.001f) cachedRayleighScale = 13.558f;
-            if (cachedMieScale < 0.001f) cachedMieScale = 3.996f;
-            atmosParams.rayleighScatteringBase = glm::vec3(0.0f);
-            atmosParams.mieScatteringBase = 0.0f;
-            atmosParams.mieAbsorptionBase = 0.0f;
-            atmosParams.ozoneAbsorption = glm::vec3(0.0f);
-            atmosChanged = true;
-        }
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Toggle sky scattering (Rayleigh blue sky, Mie haze)");
-    }
-
-    if (atmosphereEnabled) {
-        // Rayleigh scattering (blue sky) - wide ranges for extreme testing
-        ImGui::Text("Rayleigh Scattering (Air):");
-        float rayleighScale = atmosParams.rayleighScatteringBase.y * 1000.0f;  // Scale for UI
-        if (ImGui::SliderFloat("Rayleigh Strength", &rayleighScale, 0.0f, 200.0f, "%.2f", ImGuiSliderFlags_Logarithmic)) {
-            float oldVal = atmosParams.rayleighScatteringBase.y * 1000.0f;
-            if (oldVal > 0.0001f) {
-                float ratio = rayleighScale / oldVal;
-                atmosParams.rayleighScatteringBase *= ratio;
-            } else {
-                // If starting from near zero, set to Earth-like ratio
-                atmosParams.rayleighScatteringBase = glm::vec3(5.802e-3f, 13.558e-3f, 33.1e-3f) * (rayleighScale / 13.558f);
-            }
-            cachedRayleighScale = rayleighScale;
-            atmosChanged = true;
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0 = no blue sky, 13.5 = Earth, 200 = extremely blue (logarithmic)");
-        }
-
-        if (ImGui::SliderFloat("Rayleigh Scale Height", &atmosParams.rayleighScaleHeight, 0.1f, 100.0f, "%.1f km", ImGuiSliderFlags_Logarithmic)) {
-            atmosChanged = true;
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0.1 = thin atmosphere, 8 = Earth, 100 = very thick");
-        }
-
-        // Mie scattering (haze/sun halo) - wide ranges
-        ImGui::Spacing();
-        ImGui::Text("Mie Scattering (Haze):");
-        float mieScale = atmosParams.mieScatteringBase * 1000.0f;
-        if (ImGui::SliderFloat("Mie Strength", &mieScale, 0.0f, 200.0f, "%.2f", ImGuiSliderFlags_Logarithmic)) {
-            atmosParams.mieScatteringBase = mieScale / 1000.0f;
-            cachedMieScale = mieScale;
-            atmosChanged = true;
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0 = no haze, 4 = Earth, 200 = dense smog (logarithmic)");
-        }
-
-        if (ImGui::SliderFloat("Mie Scale Height", &atmosParams.mieScaleHeight, 0.01f, 50.0f, "%.2f km", ImGuiSliderFlags_Logarithmic)) {
-            atmosChanged = true;
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0.01 = ground-level only, 1.2 = Earth, 50 = everywhere");
-        }
-
-        if (ImGui::SliderFloat("Mie Anisotropy", &atmosParams.mieAnisotropy, -0.99f, 0.99f, "%.2f")) {
-            atmosChanged = true;
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("-1 = backward scatter, 0 = uniform, 0.8 = Earth (forward), 0.99 = laser-like sun");
-        }
-
-        float mieAbsScale = atmosParams.mieAbsorptionBase * 1000.0f;
-        if (ImGui::SliderFloat("Mie Absorption", &mieAbsScale, 0.0f, 100.0f, "%.2f", ImGuiSliderFlags_Logarithmic)) {
-            atmosParams.mieAbsorptionBase = mieAbsScale / 1000.0f;
-            atmosChanged = true;
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0 = no absorption, 4.4 = Earth, 100 = heavy smog");
-        }
-
-        // Ozone (affects horizon color) - wide ranges
-        ImGui::Spacing();
-        ImGui::Text("Ozone Layer:");
-        float ozoneScale = atmosParams.ozoneAbsorption.y * 1000.0f;
-        if (ImGui::SliderFloat("Ozone Strength", &ozoneScale, 0.0f, 50.0f, "%.2f", ImGuiSliderFlags_Logarithmic)) {
-            float oldVal = atmosParams.ozoneAbsorption.y * 1000.0f;
-            if (oldVal > 0.0001f) {
-                float ratio = ozoneScale / oldVal;
-                atmosParams.ozoneAbsorption *= ratio;
-            } else {
-                // If starting from near zero, set to Earth-like ratio
-                atmosParams.ozoneAbsorption = glm::vec3(0.65e-3f, 1.881e-3f, 0.085e-3f) * (ozoneScale / 1.881f);
-            }
-            atmosChanged = true;
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0 = no ozone, 1.9 = Earth, 50 = extreme orange sunsets");
-        }
-
-        if (ImGui::SliderFloat("Ozone Center", &atmosParams.ozoneLayerCenter, 0.0f, 100.0f, "%.0f km")) {
-            atmosChanged = true;
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0 = at surface, 25 = Earth, 100 = very high");
-        }
-
-        if (ImGui::SliderFloat("Ozone Width", &atmosParams.ozoneLayerWidth, 0.1f, 100.0f, "%.1f km", ImGuiSliderFlags_Logarithmic)) {
-            atmosChanged = true;
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("0.1 = thin band, 15 = Earth, 100 = everywhere");
-        }
-
-        // Quick presets
-        ImGui::Spacing();
-        ImGui::Text("Presets:");
-        if (ImGui::Button("Earth##atmos")) {
-            AtmosphereParams earth;
-            renderer.setAtmosphereParams(earth);
-            cachedRayleighScale = 13.558f;
-            cachedMieScale = 3.996f;
-            atmosChanged = false;  // Already set
-        }
-        ImGui::SameLine();
-        if (ImGui::Button("Clear##atmos")) {
-            AtmosphereParams clear;
-            clear.mieScatteringBase = 1.0e-3f;
-            clear.mieAbsorptionBase = 1.0e-3f;
-            renderer.setAtmosphereParams(clear);
-            cachedMieScale = 1.0f;
-            atmosChanged = false;
-        }
-        ImGui::SameLine();
-        if (ImGui::Button("Hazy##atmos")) {
-            AtmosphereParams hazy;
-            hazy.mieScatteringBase = 15.0e-3f;
-            hazy.mieAbsorptionBase = 10.0e-3f;
-            hazy.mieAnisotropy = 0.7f;
-            renderer.setAtmosphereParams(hazy);
-            cachedMieScale = 15.0f;
-            atmosChanged = false;
-        }
-    }
-
-    if (atmosChanged) {
-        renderer.setAtmosphereParams(atmosParams);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Leaf system
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.9f, 0.7f, 0.5f, 1.0f));
-    ImGui::Text("FALLING LEAVES");
-    ImGui::PopStyleColor();
-
-    float leafIntensity = renderer.getLeafIntensity();
-    if (ImGui::SliderFloat("Leaf Intensity", &leafIntensity, 0.0f, 1.0f)) {
-        renderer.setLeafIntensity(leafIntensity);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Cloud style
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.9f, 0.9f, 0.7f, 1.0f));
-    ImGui::Text("CLOUDS");
-    ImGui::PopStyleColor();
-
-    bool paraboloid = renderer.isUsingParaboloidClouds();
-    if (ImGui::Checkbox("Paraboloid LUT Clouds", &paraboloid)) {
-        renderer.toggleCloudStyle();
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Toggle between procedural and paraboloid LUT hybrid cloud rendering");
-    }
-
-    // Cloud coverage and density controls
-    float cloudCoverage = renderer.getCloudCoverage();
-    if (ImGui::SliderFloat("Cloud Coverage", &cloudCoverage, 0.0f, 1.0f, "%.2f")) {
-        renderer.setCloudCoverage(cloudCoverage);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("0 = clear sky, 0.5 = partly cloudy, 1 = overcast");
-    }
-
-    float cloudDensity = renderer.getCloudDensity();
-    if (ImGui::SliderFloat("Cloud Density", &cloudDensity, 0.0f, 1.0f, "%.2f")) {
-        renderer.setCloudDensity(cloudDensity);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("0 = thin/wispy, 0.3 = normal, 1 = thick/opaque");
-    }
-
-    // Cloud presets
-    ImGui::Text("Presets:");
-    ImGui::SameLine();
-    if (ImGui::Button("Clear##clouds")) {
-        renderer.setCloudCoverage(0.0f);
-        renderer.setCloudDensity(0.3f);
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Partly##clouds")) {
-        renderer.setCloudCoverage(0.4f);
-        renderer.setCloudDensity(0.3f);
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Cloudy##clouds")) {
-        renderer.setCloudCoverage(0.7f);
-        renderer.setCloudDensity(0.5f);
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Overcast##clouds")) {
-        renderer.setCloudCoverage(0.95f);
-        renderer.setCloudDensity(0.7f);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Grass interaction
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.5f, 0.9f, 0.5f, 1.0f));
-    ImGui::Text("GRASS INTERACTION");
-    ImGui::PopStyleColor();
-
-    auto& env = renderer.getEnvironmentSettings();
-    if (ImGui::SliderFloat("Displacement Decay", &env.grassDisplacementDecay, 0.1f, 5.0f)) {}
-    if (ImGui::SliderFloat("Max Displacement", &env.grassMaxDisplacement, 0.0f, 2.0f)) {}
-}
-
-void GuiSystem::renderPostProcessSection(Renderer& renderer) {
-    ImGui::Spacing();
-
-    // HDR Tonemapping toggle
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.7f, 0.4f, 1.0f));
-    ImGui::Text("HDR PIPELINE");
-    ImGui::PopStyleColor();
-
-    bool hdrEnabled = renderer.isHDREnabled();
-    if (ImGui::Checkbox("HDR Tonemapping", &hdrEnabled)) {
-        renderer.setHDREnabled(hdrEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Enable/disable ACES tonemapping and exposure control");
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Cloud shadows toggle
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.8f, 1.0f, 1.0f));
-    ImGui::Text("CLOUD SHADOWS");
-    ImGui::PopStyleColor();
-
-    bool cloudShadowEnabled = renderer.isCloudShadowEnabled();
-    if (ImGui::Checkbox("Cloud Shadows", &cloudShadowEnabled)) {
-        renderer.setCloudShadowEnabled(cloudShadowEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Enable/disable cloud shadow projection on terrain");
-    }
-
-    if (cloudShadowEnabled) {
-        float cloudShadowIntensity = renderer.getCloudShadowIntensity();
-        if (ImGui::SliderFloat("Shadow Intensity", &cloudShadowIntensity, 0.0f, 1.0f)) {
-            renderer.setCloudShadowIntensity(cloudShadowIntensity);
-        }
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.8f, 0.5f, 1.0f));
-    ImGui::Text("BLOOM");
-    ImGui::PopStyleColor();
-
-    ImGui::TextDisabled("Bloom is enabled by default");
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.9f, 0.6f, 1.0f));
-    ImGui::Text("GOD RAYS");
-    ImGui::PopStyleColor();
-
-    bool godRaysEnabled = renderer.isGodRaysEnabled();
-    if (ImGui::Checkbox("Enable God Rays", &godRaysEnabled)) {
-        renderer.setGodRaysEnabled(godRaysEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Toggle god ray light shafts effect");
-    }
-
-    if (godRaysEnabled) {
-        // God ray quality dropdown
-        const char* qualityNames[] = {"Low (16 samples)", "Medium (32 samples)", "High (64 samples)"};
-        int currentQuality = static_cast<int>(renderer.getGodRayQuality());
-        if (ImGui::Combo("God Ray Quality", &currentQuality, qualityNames, 3)) {
-            renderer.setGodRayQuality(static_cast<PostProcessSystem::GodRayQuality>(currentQuality));
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("Higher quality = more samples = better rays but slower");
-        }
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.9f, 1.0f, 1.0f));
-    ImGui::Text("VOLUMETRIC FOG");
-    ImGui::PopStyleColor();
-
-    bool froxelHighQuality = renderer.isFroxelFilterHighQuality();
-    if (ImGui::Checkbox("High Quality Fog Filter", &froxelHighQuality)) {
-        renderer.setFroxelFilterQuality(froxelHighQuality);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Tricubic filtering (8 samples) vs Trilinear (1 sample)");
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.8f, 0.8f, 1.0f, 1.0f));
-    ImGui::Text("EXPOSURE");
-    ImGui::PopStyleColor();
-
-    ImGui::TextDisabled("Auto-exposure is active");
-    ImGui::TextDisabled("Histogram-based adaptation");
-}
-
-void GuiSystem::renderTerrainSection(Renderer& renderer) {
-    ImGui::Spacing();
-
-    // Terrain info
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.6f, 0.8f, 0.6f, 1.0f));
-    ImGui::Text("TERRAIN SYSTEM");
-    ImGui::PopStyleColor();
-
-    const auto& terrain = renderer.getTerrainSystem();
-    const auto& config = terrain.getConfig();
-
-    ImGui::Text("Size: %.0f x %.0f meters", config.size, config.size);
-    ImGui::Text("Height Scale: %.1f", config.heightScale);
-
-    // Triangle count with color coding
-    uint32_t triangleCount = renderer.getTerrainNodeCount();
-    ImVec4 triColor = triangleCount < 100000 ? ImVec4(0.4f, 0.9f, 0.4f, 1.0f) :
-                      triangleCount < 500000 ? ImVec4(0.9f, 0.9f, 0.4f, 1.0f) :
-                                               ImVec4(0.9f, 0.4f, 0.4f, 1.0f);
-    ImGui::Text("Triangles:");
-    ImGui::SameLine();
-    ImGui::PushStyleColor(ImGuiCol_Text, triColor);
-    if (triangleCount >= 1000000) {
-        ImGui::Text("%.2fM", triangleCount / 1000000.0f);
-    } else if (triangleCount >= 1000) {
-        ImGui::Text("%.1fK", triangleCount / 1000.0f);
-    } else {
-        ImGui::Text("%u", triangleCount);
-    }
-    ImGui::PopStyleColor();
-
-    // CBT depth info
-    ImGui::Text("Max Depth: %d (min edge: %.1fm)", config.maxDepth,
-                config.size / (1 << (config.maxDepth / 2)));
-    ImGui::Text("Min Depth: %d", config.minDepth);
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // LOD parameters (modifiable at runtime)
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.8f, 0.7f, 0.5f, 1.0f));
-    ImGui::Text("LOD PARAMETERS");
-    ImGui::PopStyleColor();
-
-    auto& terrainMut = renderer.getTerrainSystem();
-    TerrainConfig cfg = terrainMut.getConfig();
-    bool configChanged = false;
-
-    if (ImGui::SliderFloat("Split Threshold", &cfg.splitThreshold, 1.0f, 256.0f, "%.0f px")) {
-        configChanged = true;
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Screen-space edge length (pixels) to trigger subdivision");
-    }
-
-    if (ImGui::SliderFloat("Merge Threshold", &cfg.mergeThreshold, 1.0f, 256.0f, "%.0f px")) {
-        configChanged = true;
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Screen-space edge length (pixels) to trigger merge");
-    }
-
-    if (ImGui::SliderFloat("Flatness Scale", &cfg.flatnessScale, 0.0f, 5.0f, "%.1f")) {
-        configChanged = true;
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Curvature LOD: 0=disabled, 2=flat areas use 3x threshold");
-    }
-
-    int maxDepth = cfg.maxDepth;
-    if (ImGui::SliderInt("Max Depth", &maxDepth, 16, 32)) {
-        cfg.maxDepth = maxDepth;
-        configChanged = true;
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Maximum subdivision depth (higher = finer detail, more triangles)");
-    }
-
-    int minDepth = cfg.minDepth;
-    if (ImGui::SliderInt("Min Depth", &minDepth, 1, 10)) {
-        cfg.minDepth = minDepth;
-        configChanged = true;
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Minimum subdivision depth (base tessellation level)");
-    }
-
-    int spreadFactor = static_cast<int>(cfg.spreadFactor);
-    if (ImGui::SliderInt("Spread Factor", &spreadFactor, 1, 32)) {
-        cfg.spreadFactor = static_cast<uint32_t>(spreadFactor);
-        configChanged = true;
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Temporal spreading: process 1/N triangles per frame (1 = all, higher = less GPU work per frame)");
-    }
-
-    if (configChanged) {
-        terrainMut.setConfig(cfg);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Debug toggles
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.9f, 0.6f, 0.6f, 1.0f));
-    ImGui::Text("DEBUG");
-    ImGui::PopStyleColor();
-
-    bool terrainEnabled = renderer.isTerrainEnabled();
-    if (ImGui::Checkbox("Enable Terrain", &terrainEnabled)) {
-        renderer.setTerrainEnabled(terrainEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Toggle terrain rendering on/off");
-    }
-
-    bool wireframe = renderer.isTerrainWireframeMode();
-    if (ImGui::Checkbox("Wireframe Mode", &wireframe)) {
-        renderer.toggleTerrainWireframe();
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Show terrain mesh wireframe overlay");
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Meshlet rendering
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.5f, 0.8f, 0.9f, 1.0f));
-    ImGui::Text("MESHLET RENDERING");
-    ImGui::PopStyleColor();
-
-    bool meshletsEnabled = terrainMut.isMeshletsEnabled();
-    if (ImGui::Checkbox("Enable Meshlets", &meshletsEnabled)) {
-        terrainMut.setMeshletsEnabled(meshletsEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Use pre-tessellated meshlets per CBT leaf for higher resolution");
-    }
-
-    if (meshletsEnabled) {
-        int meshletLevel = terrainMut.getMeshletSubdivisionLevel();
-        if (ImGui::SliderInt("Meshlet Level", &meshletLevel, 0, 6)) {
-            terrainMut.setMeshletSubdivisionLevel(meshletLevel);
-        }
-        if (ImGui::IsItemHovered()) {
-            ImGui::SetTooltip("Subdivision level per meshlet (0=1, 1=4, 2=16, 3=64, 4=256 triangles)");
-        }
-
-        uint32_t meshletTris = terrainMut.getMeshletTriangleCount();
-        ImGui::Text("Triangles per leaf: %u", meshletTris);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Optimization toggles
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.8f, 0.6f, 0.8f, 1.0f));
-    ImGui::Text("OPTIMIZATIONS");
-    ImGui::PopStyleColor();
-
-    bool skipFrameOpt = terrainMut.isSkipFrameOptimizationEnabled();
-    if (ImGui::Checkbox("Skip-Frame (Camera Still)", &skipFrameOpt)) {
-        terrainMut.setSkipFrameOptimization(skipFrameOpt);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Skip subdivision compute when camera is stationary");
-    }
-
-    bool gpuCulling = terrainMut.isGpuCullingEnabled();
-    if (ImGui::Checkbox("GPU Frustum Culling", &gpuCulling)) {
-        terrainMut.setGpuCulling(gpuCulling);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Use GPU frustum culling with stream compaction for split phase");
-    }
-
-    // Debug status (phase alternates every frame so not useful to display)
-    ImGui::Text("Status: %s", terrainMut.isCurrentlySkipping() ? "SKIPPING" : "ACTIVE");
-
-    ImGui::Spacing();
-
-    // Height query demo
-    ImGui::Text("Height at origin: %.2f", renderer.getTerrainHeightAt(0.0f, 0.0f));
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Streaming stats (Ghost of Tsushima style)
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.6f, 0.9f, 0.8f, 1.0f));
-    ImGui::Text("STREAMING");
-    ImGui::PopStyleColor();
-
-    const auto& tileCache = terrain.getTileCache();
-    uint32_t activeTiles = tileCache.getActiveTileCount();
-    uint32_t maxTiles = 64;  // MAX_ACTIVE_TILES constant
-
-    // Active tiles with color coding
-    float tileUsage = static_cast<float>(activeTiles) / static_cast<float>(maxTiles);
-    ImVec4 tileColor = tileUsage < 0.5f ? ImVec4(0.4f, 0.9f, 0.4f, 1.0f) :
-                       tileUsage < 0.8f ? ImVec4(0.9f, 0.9f, 0.4f, 1.0f) :
-                                          ImVec4(0.9f, 0.4f, 0.4f, 1.0f);
-    ImGui::Text("Height Tiles:");
-    ImGui::SameLine();
-    ImGui::PushStyleColor(ImGuiCol_Text, tileColor);
-    ImGui::Text("%u / %u (%.0f%%)", activeTiles, maxTiles, tileUsage * 100.0f);
-    ImGui::PopStyleColor();
-
-    // LOD info
-    ImGui::Text("LOD Ranges: <1km:LOD0, 1-2km:LOD1, 2-4km:LOD2");
-
-    ImGui::Spacing();
-
-    // Render distance controls
-    TerrainConfig streamCfg = terrainMut.getConfig();
-    bool streamConfigChanged = false;
-
-    if (ImGui::SliderFloat("Load Radius", &streamCfg.tileLoadRadius, 500.0f, 8000.0f, "%.0f m")) {
-        // Ensure unload radius stays larger than load radius
-        if (streamCfg.tileUnloadRadius < streamCfg.tileLoadRadius + 500.0f) {
-            streamCfg.tileUnloadRadius = streamCfg.tileLoadRadius + 500.0f;
-        }
-        streamConfigChanged = true;
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Distance from camera to load high-resolution height tiles");
-    }
-
-    if (ImGui::SliderFloat("Unload Radius", &streamCfg.tileUnloadRadius, 1000.0f, 10000.0f, "%.0f m")) {
-        // Ensure unload radius stays larger than load radius
-        if (streamCfg.tileUnloadRadius < streamCfg.tileLoadRadius + 500.0f) {
-            streamCfg.tileUnloadRadius = streamCfg.tileLoadRadius + 500.0f;
-        }
-        streamConfigChanged = true;
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Distance from camera to unload tiles (should be > load radius)");
-    }
-
-    if (streamConfigChanged) {
-        terrainMut.setConfig(streamCfg);
-    }
-}
-
-void GuiSystem::renderWaterSection(Renderer& renderer) {
-    ImGui::Spacing();
-
-    auto& water = renderer.getWaterSystem();
-
-    // Water info header
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.4f, 0.7f, 0.9f, 1.0f));
-    ImGui::Text("WATER SYSTEM");
-    ImGui::PopStyleColor();
-
-    ImGui::Text("Current Level: %.2f m", water.getWaterLevel());
-    ImGui::Text("Base Level: %.2f m", water.getBaseWaterLevel());
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Water level controls
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.5f, 0.8f, 1.0f, 1.0f));
-    ImGui::Text("LEVEL & TIDES");
-    ImGui::PopStyleColor();
-
-    float baseLevel = water.getBaseWaterLevel();
-    if (ImGui::SliderFloat("Base Water Level", &baseLevel, -50.0f, 50.0f, "%.1f m")) {
-        water.setWaterLevel(baseLevel);
-    }
-
-    float tidalRange = water.getTidalRange();
-    if (ImGui::SliderFloat("Tidal Range", &tidalRange, 0.0f, 10.0f, "%.1f m")) {
-        water.setTidalRange(tidalRange);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Maximum tide variation from base level");
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Wave parameters
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.6f, 0.9f, 0.8f, 1.0f));
-    ImGui::Text("WAVES");
-    ImGui::PopStyleColor();
-
-    // FFT Ocean toggle
-    bool useFFT = water.getUseFFTOcean();
-    if (ImGui::Checkbox("FFT Ocean (Tessendorf)", &useFFT)) {
-        water.setUseFFTOcean(useFFT);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Use FFT-based ocean simulation instead of Gerstner waves");
-    }
-
-    float amplitude = water.getWaveAmplitude();
-    if (ImGui::SliderFloat("Amplitude", &amplitude, 0.0f, 5.0f, "%.2f m")) {
-        water.setWaveAmplitude(amplitude);
-    }
-
-    float wavelength = water.getWaveLength();
-    if (ImGui::SliderFloat("Wavelength", &wavelength, 1.0f, 100.0f, "%.1f m")) {
-        water.setWaveLength(wavelength);
-    }
-
-    float steepness = water.getWaveSteepness();
-    if (ImGui::SliderFloat("Steepness", &steepness, 0.0f, 1.0f, "%.2f")) {
-        water.setWaveSteepness(steepness);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Wave sharpness (0=sine, 1=peaked)");
-    }
-
-    float speed = water.getWaveSpeed();
-    if (ImGui::SliderFloat("Speed", &speed, 0.0f, 3.0f, "%.2f")) {
-        water.setWaveSpeed(speed);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Appearance
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.8f, 1.0f, 1.0f));
-    ImGui::Text("APPEARANCE");
-    ImGui::PopStyleColor();
-
-    glm::vec4 waterColor = water.getWaterColor();
-    float col[4] = {waterColor.r, waterColor.g, waterColor.b, waterColor.a};
-    if (ImGui::ColorEdit4("Water Color", col)) {
-        water.setWaterColor(glm::vec4(col[0], col[1], col[2], col[3]));
-    }
-
-    float foam = water.getFoamThreshold();
-    if (ImGui::SliderFloat("Foam Threshold", &foam, 0.0f, 2.0f, "%.2f")) {
-        water.setFoamThreshold(foam);
-    }
-
-    float fresnel = water.getFresnelPower();
-    if (ImGui::SliderFloat("Fresnel Power", &fresnel, 1.0f, 10.0f, "%.1f")) {
-        water.setFresnelPower(fresnel);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Controls reflection intensity at grazing angles");
-    }
-
-    // Shore effects
-    ImGui::Spacing();
-    ImGui::Text("Shore Effects:");
-
-    float shoreBlend = water.getShoreBlendDistance();
-    if (ImGui::SliderFloat("Shore Blend", &shoreBlend, 0.5f, 10.0f, "%.1f m")) {
-        water.setShoreBlendDistance(shoreBlend);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Distance over which water fades near shore");
-    }
-
-    float shoreFoam = water.getShoreFoamWidth();
-    if (ImGui::SliderFloat("Shore Foam Width", &shoreFoam, 1.0f, 20.0f, "%.1f m")) {
-        water.setShoreFoamWidth(shoreFoam);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Width of foam bands along the shoreline");
-    }
-
-    // Presets
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    ImGui::Text("Presets:");
-    if (ImGui::Button("Ocean")) {
-        water.setWaterColor(glm::vec4(0.02f, 0.08f, 0.15f, 0.95f));
-        water.setWaveAmplitude(1.5f);
-        water.setWaveLength(30.0f);
-        water.setWaveSteepness(0.4f);
-        water.setWaveSpeed(0.8f);
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Lake")) {
-        water.setWaterColor(glm::vec4(0.05f, 0.12f, 0.18f, 0.9f));
-        water.setWaveAmplitude(0.3f);
-        water.setWaveLength(8.0f);
-        water.setWaveSteepness(0.2f);
-        water.setWaveSpeed(0.5f);
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Calm")) {
-        water.setWaterColor(glm::vec4(0.03f, 0.1f, 0.2f, 0.85f));
-        water.setWaveAmplitude(0.1f);
-        water.setWaveLength(5.0f);
-        water.setWaveSteepness(0.1f);
-        water.setWaveSpeed(0.3f);
-    }
-    if (ImGui::Button("Storm")) {
-        water.setWaterColor(glm::vec4(0.04f, 0.06f, 0.1f, 0.98f));
-        water.setWaveAmplitude(3.0f);
-        water.setWaveLength(20.0f);
-        water.setWaveSteepness(0.6f);
-        water.setWaveSpeed(1.5f);
-    }
-    ImGui::SameLine();
-    if (ImGui::Button("Tropical")) {
-        water.setWaterColor(glm::vec4(0.0f, 0.15f, 0.2f, 0.8f));
-        water.setWaveAmplitude(0.5f);
-        water.setWaveLength(12.0f);
-        water.setWaveSteepness(0.3f);
-        water.setWaveSpeed(0.6f);
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Performance optimization controls
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.9f, 0.7f, 0.5f, 1.0f));
-    ImGui::Text("PERFORMANCE");
-    ImGui::PopStyleColor();
-
-    auto& tileCull = renderer.getWaterTileCull();
-    bool tileCullEnabled = tileCull.isEnabled();
-    if (ImGui::Checkbox("Tile Culling", &tileCullEnabled)) {
-        tileCull.setEnabled(tileCullEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Skip water rendering when not visible (temporal)");
-    }
-
-    // Show tile cull stats when enabled
-    if (tileCullEnabled) {
-        glm::uvec2 tileCount = tileCull.getTileCount();
-        ImGui::Text("Tiles: %ux%u (%u px)", tileCount.x, tileCount.y, tileCull.getTileSize());
-    }
-}
-
-void GuiSystem::renderDebugSection(Renderer& renderer) {
-    ImGui::Spacing();
-
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.6f, 0.6f, 1.0f));
-    ImGui::Text("DEBUG VISUALIZATIONS");
-    ImGui::PopStyleColor();
-
-    bool cascadeDebug = renderer.isShowingCascadeDebug();
-    if (ImGui::Checkbox("Shadow Cascade Debug", &cascadeDebug)) {
-        renderer.toggleCascadeDebug();
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Shows colored overlay for each shadow cascade");
-    }
-
-    bool snowDepthDebug = renderer.isShowingSnowDepthDebug();
-    if (ImGui::Checkbox("Snow Depth Debug", &snowDepthDebug)) {
-        renderer.toggleSnowDepthDebug();
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Shows snow accumulation depth as heat map");
-    }
-
-#ifdef JPH_DEBUG_RENDERER
-    ImGui::Spacing();
-
-    bool physicsDebug = renderer.isPhysicsDebugEnabled();
-    if (ImGui::Checkbox("Physics Debug", &physicsDebug)) {
-        renderer.setPhysicsDebugEnabled(physicsDebug);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Draw Jolt Physics collision shapes and debug info");
-    }
-
-    if (physicsDebug) {
-        ImGui::Indent();
-
-        auto* debugRenderer = renderer.getPhysicsDebugRenderer();
-        if (debugRenderer) {
-            auto& options = debugRenderer->getOptions();
-
-            ImGui::Checkbox("Draw Shapes", &options.drawShapes);
-            ImGui::Checkbox("Wireframe", &options.drawShapeWireframe);
-            ImGui::Checkbox("Bounding Boxes", &options.drawBoundingBox);
-            ImGui::Checkbox("Velocity", &options.drawVelocity);
-            ImGui::Checkbox("Center of Mass", &options.drawCenterOfMassTransform);
-
-            ImGui::Spacing();
-            ImGui::Text("Body Types:");
-            ImGui::Checkbox("Static", &options.drawStaticBodies);
-            if (ImGui::IsItemHovered()) {
-                ImGui::SetTooltip("Warning: Static bodies include terrain heightfields which are very slow to render");
-            }
-            ImGui::Checkbox("Dynamic", &options.drawDynamicBodies);
-            ImGui::Checkbox("Kinematic", &options.drawKinematicBodies);
-            ImGui::Checkbox("Character", &options.drawCharacter);
-        } else {
-            ImGui::TextDisabled("Enable to see options");
-        }
-
-        ImGui::Unindent();
-    }
-
-    // Show stats
-    auto& debugLines = renderer.getDebugLineSystem();
-    if (physicsDebug) {
-        ImGui::Spacing();
-        ImGui::Text("Lines: %zu", debugLines.getLineCount());
-        ImGui::Text("Triangles: %zu", debugLines.getTriangleCount());
-    }
-#endif
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.6f, 0.8f, 1.0f, 1.0f));
-    ImGui::Text("OCCLUSION CULLING");
-    ImGui::PopStyleColor();
-
-    bool hiZEnabled = renderer.isHiZCullingEnabled();
-    if (ImGui::Checkbox("Hi-Z Occlusion Culling", &hiZEnabled)) {
-        renderer.setHiZCullingEnabled(hiZEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Enable/disable hierarchical Z-buffer occlusion culling (8 key)");
-    }
-
-    // Display culling statistics
-    auto stats = renderer.getHiZCullingStats();
-    ImGui::Text("Total Objects: %u", stats.totalObjects);
-    ImGui::Text("Visible: %u", stats.visibleObjects);
-    ImGui::Text("Frustum Culled: %u", stats.frustumCulled);
-    ImGui::Text("Occlusion Culled: %u", stats.occlusionCulled);
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.8f, 0.8f, 0.5f, 1.0f));
-    ImGui::Text("SYSTEM INFO");
-    ImGui::PopStyleColor();
-
-    ImGui::Text("Renderer: Vulkan");
-    ImGui::Text("Shadow Cascades: 4");
-    ImGui::Text("Shadow Map Size: 2048");
-    ImGui::Text("Max Frames in Flight: 2");
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Keyboard shortcuts reference
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.7f, 0.7f, 1.0f));
-    ImGui::Text("KEYBOARD SHORTCUTS");
-    ImGui::PopStyleColor();
-
-    ImGui::BulletText("F1 - Toggle GUI");
-    ImGui::BulletText("F2 - Tree Editor");
-    ImGui::BulletText("P - Place tree at camera");
-    ImGui::BulletText("Tab - Toggle camera mode");
-    ImGui::BulletText("1-4 - Time presets");
-    ImGui::BulletText("+/- - Time scale");
-    ImGui::BulletText("C - Cycle weather");
-    ImGui::BulletText("Z/X - Weather intensity");
-    ImGui::BulletText(",/. - Snow amount");
-    ImGui::BulletText("T - Terrain wireframe");
-    ImGui::BulletText("6 - Cascade debug");
-    ImGui::BulletText("7 - Snow depth debug");
-    ImGui::BulletText("8 - Hi-Z culling toggle");
-    ImGui::BulletText("[ ] - Fog density");
-    ImGui::BulletText("\\ - Toggle fog");
-    ImGui::BulletText("F - Spawn confetti");
-}
-
-void GuiSystem::renderProfilerSection(Renderer& renderer) {
-    ImGui::Spacing();
-
-    auto& profiler = renderer.getProfiler();
-
-    // Enable/disable toggle
-    bool enabled = profiler.isEnabled();
-    if (ImGui::Checkbox("Enable Profiling", &enabled)) {
-        profiler.setEnabled(enabled);
-    }
-
-    if (!enabled) {
-        ImGui::TextDisabled("Profiling disabled");
-        return;
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // GPU Profiling Section
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.4f, 0.8f, 1.0f, 1.0f));
-    ImGui::Text("GPU TIMING");
-    ImGui::PopStyleColor();
-
-    const auto& gpuStats = profiler.getSmoothedGpuResults();
-
-    if (gpuStats.zones.empty()) {
-        ImGui::TextDisabled("No GPU data yet (waiting for frames)");
-    } else {
-        // Total GPU time
-        ImGui::Text("Total GPU: %.2f ms", gpuStats.totalGpuTimeMs);
-
-        ImGui::Spacing();
-
-        // GPU timing breakdown table
-        if (ImGui::BeginTable("GPUTimings", 3, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
-            ImGui::TableSetupColumn("Pass", ImGuiTableColumnFlags_WidthStretch);
-            ImGui::TableSetupColumn("Time (ms)", ImGuiTableColumnFlags_WidthFixed, 70.0f);
-            ImGui::TableSetupColumn("%", ImGuiTableColumnFlags_WidthFixed, 50.0f);
-            ImGui::TableHeadersRow();
-
-            for (const auto& zone : gpuStats.zones) {
-                ImGui::TableNextRow();
-
-                ImGui::TableNextColumn();
-                ImGui::Text("%s", zone.name.c_str());
-
-                ImGui::TableNextColumn();
-                ImGui::Text("%.2f", zone.gpuTimeMs);
-
-                ImGui::TableNextColumn();
-                // Color code by percentage
-                if (zone.percentOfFrame > 30.0f) {
-                    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.4f, 0.4f, 1.0f));
-                } else if (zone.percentOfFrame > 15.0f) {
-                    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.8f, 0.4f, 1.0f));
-                } else {
-                    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.4f, 1.0f, 0.4f, 1.0f));
-                }
-                ImGui::Text("%.1f%%", zone.percentOfFrame);
-                ImGui::PopStyleColor();
-            }
-
-            ImGui::EndTable();
-        }
-
-        // Visual bar chart of GPU zones
-        ImGui::Spacing();
-        float maxTime = gpuStats.totalGpuTimeMs;
-        for (const auto& zone : gpuStats.zones) {
-            float fraction = (maxTime > 0.0f) ? (zone.gpuTimeMs / maxTime) : 0.0f;
-            ImGui::ProgressBar(fraction, ImVec2(-1, 0), zone.name.c_str());
-        }
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // CPU Profiling Section
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.8f, 0.4f, 1.0f));
-    ImGui::Text("CPU TIMING");
-    ImGui::PopStyleColor();
-
-    const auto& cpuStats = profiler.getSmoothedCpuResults();
-
-    if (cpuStats.zones.empty()) {
-        ImGui::TextDisabled("No CPU data yet");
-    } else {
-        // Total CPU time
-        ImGui::Text("Total CPU: %.2f ms", cpuStats.totalCpuTimeMs);
-
-        ImGui::Spacing();
-
-        // CPU timing breakdown table
-        if (ImGui::BeginTable("CPUTimings", 3, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
-            ImGui::TableSetupColumn("Zone", ImGuiTableColumnFlags_WidthStretch);
-            ImGui::TableSetupColumn("Time (ms)", ImGuiTableColumnFlags_WidthFixed, 70.0f);
-            ImGui::TableSetupColumn("%", ImGuiTableColumnFlags_WidthFixed, 50.0f);
-            ImGui::TableHeadersRow();
-
-            for (const auto& zone : cpuStats.zones) {
-                ImGui::TableNextRow();
-
-                ImGui::TableNextColumn();
-                ImGui::Text("%s", zone.name.c_str());
-
-                ImGui::TableNextColumn();
-                ImGui::Text("%.3f", zone.cpuTimeMs);
-
-                ImGui::TableNextColumn();
-                ImGui::Text("%.1f%%", zone.percentOfFrame);
-            }
-
-            ImGui::EndTable();
-        }
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Frame budget indicator
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.8f, 0.8f, 1.0f, 1.0f));
-    ImGui::Text("FRAME BUDGET");
-    ImGui::PopStyleColor();
-
-    float targetMs = 16.67f;  // 60 FPS target
-    float gpuTime = gpuStats.totalGpuTimeMs;
-    float cpuTime = cpuStats.totalCpuTimeMs;
-    float maxTime = std::max(gpuTime, cpuTime);
-
-    // Budget bar
-    float budgetUsed = maxTime / targetMs;
-    ImVec4 budgetColor;
-    if (budgetUsed < 0.8f) {
-        budgetColor = ImVec4(0.4f, 1.0f, 0.4f, 1.0f);  // Green
-    } else if (budgetUsed < 1.0f) {
-        budgetColor = ImVec4(1.0f, 0.8f, 0.4f, 1.0f);  // Yellow
-    } else {
-        budgetColor = ImVec4(1.0f, 0.4f, 0.4f, 1.0f);  // Red
-    }
-
-    ImGui::PushStyleColor(ImGuiCol_PlotHistogram, budgetColor);
-    char budgetText[64];
-    snprintf(budgetText, sizeof(budgetText), "%.1f / %.1f ms (%.0f%%)",
-             maxTime, targetMs, budgetUsed * 100.0f);
-    ImGui::ProgressBar(std::min(budgetUsed, 1.5f) / 1.5f, ImVec2(-1, 20), budgetText);
-    ImGui::PopStyleColor();
-
-    ImGui::Text("GPU Bound: %s", (gpuTime > cpuTime) ? "Yes" : "No");
-    ImGui::Text("CPU Bound: %s", (cpuTime > gpuTime) ? "Yes" : "No");
 }
 
 void GuiSystem::renderHelpOverlay() {
@@ -2136,435 +572,4 @@ void GuiSystem::renderPositionPanel(const Camera& camera) {
         ImGui::Text("Bearing: %.0f", bearing);
     }
     ImGui::End();
-}
-
-void GuiSystem::renderIKSection(Renderer& renderer, const Camera& camera) {
-    ImGui::Spacing();
-
-    // Check if character is loaded
-    auto& sceneBuilder = renderer.getSceneBuilder();
-    if (!sceneBuilder.hasCharacter()) {
-        ImGui::TextDisabled("No animated character loaded");
-        return;
-    }
-
-    auto& character = sceneBuilder.getAnimatedCharacter();
-    auto& ikSystem = character.getIKSystem();
-
-    // Debug Visualization section
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.5f, 0.9f, 0.5f, 1.0f));
-    ImGui::Text("DEBUG VISUALIZATION");
-    ImGui::PopStyleColor();
-
-    ImGui::Checkbox("Show Skeleton", &ikDebugSettings.showSkeleton);
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Draw wireframe skeleton bones");
-    }
-
-    ImGui::Checkbox("Show IK Targets", &ikDebugSettings.showIKTargets);
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Draw IK target positions and pole vectors");
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Look-At IK section
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.5f, 0.8f, 1.0f, 1.0f));
-    ImGui::Text("LOOK-AT IK");
-    ImGui::PopStyleColor();
-
-    if (ImGui::Checkbox("Enable Look-At", &ikDebugSettings.lookAtEnabled)) {
-        ikSystem.setLookAtEnabled(ikDebugSettings.lookAtEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Character head/neck tracks a target");
-    }
-
-    if (ikDebugSettings.lookAtEnabled) {
-        ImGui::Indent();
-
-        const char* modeNames[] = { "Fixed Point", "Camera Position", "Mouse (Screen)" };
-        int currentMode = static_cast<int>(ikDebugSettings.lookAtMode);
-        if (ImGui::Combo("Target Mode", &currentMode, modeNames, 3)) {
-            ikDebugSettings.lookAtMode = static_cast<IKDebugSettings::LookAtMode>(currentMode);
-        }
-
-        if (ikDebugSettings.lookAtMode == IKDebugSettings::LookAtMode::Fixed) {
-            ImGui::DragFloat3("Target Position", &ikDebugSettings.fixedLookAtTarget.x, 0.1f);
-        }
-
-        // Update look-at target based on mode
-        glm::vec3 lookTarget;
-        switch (ikDebugSettings.lookAtMode) {
-            case IKDebugSettings::LookAtMode::Fixed:
-                lookTarget = ikDebugSettings.fixedLookAtTarget;
-                break;
-            case IKDebugSettings::LookAtMode::Camera:
-                lookTarget = camera.getPosition();
-                break;
-            case IKDebugSettings::LookAtMode::Mouse:
-                // Project mouse into world (simplified - uses camera front direction)
-                lookTarget = camera.getPosition() + camera.getFront() * 5.0f;
-                break;
-        }
-        ikSystem.setLookAtTarget(lookTarget);
-
-        float lookAtWeight = 1.0f;
-        if (ImGui::SliderFloat("Weight", &lookAtWeight, 0.0f, 1.0f)) {
-            ikSystem.setLookAtWeight(lookAtWeight);
-        }
-
-        ImGui::Unindent();
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Foot Placement IK section
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.8f, 0.5f, 1.0f));
-    ImGui::Text("FOOT PLACEMENT IK");
-    ImGui::PopStyleColor();
-
-    if (ImGui::Checkbox("Enable Foot Placement", &ikDebugSettings.footPlacementEnabled)) {
-        // Enable/disable both feet
-        ikSystem.setFootPlacementEnabled("LeftFoot", ikDebugSettings.footPlacementEnabled);
-        ikSystem.setFootPlacementEnabled("RightFoot", ikDebugSettings.footPlacementEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Feet adapt to terrain height");
-    }
-
-    if (ikDebugSettings.footPlacementEnabled) {
-        ImGui::Indent();
-
-        if (ImGui::SliderFloat("Ground Offset", &ikDebugSettings.groundOffset, -0.2f, 0.2f)) {
-            // Ground offset adjustment
-        }
-
-        // Show foot placement debug info
-        ImGui::TextDisabled("Left Foot: Active");
-        ImGui::TextDisabled("Right Foot: Active");
-
-        ImGui::Unindent();
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Straddle IK section
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.6f, 0.8f, 1.0f));
-    ImGui::Text("STRADDLE IK");
-    ImGui::PopStyleColor();
-
-    if (ImGui::Checkbox("Enable Straddling", &ikDebugSettings.straddleEnabled)) {
-        ikSystem.setStraddleEnabled(ikDebugSettings.straddleEnabled);
-    }
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Hip tilt when feet at different heights");
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // IK Chain Info
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.7f, 0.7f, 1.0f));
-    ImGui::Text("IK CHAINS");
-    ImGui::PopStyleColor();
-
-    // Show configured chains
-    const auto& skeleton = character.getSkeleton();
-    ImGui::Text("Skeleton Bones: %zu", skeleton.joints.size());
-
-    // Two-bone chains
-    if (auto* chain = ikSystem.getChain("LeftArm")) {
-        ImGui::BulletText("Left Arm: %s", chain->enabled ? "Enabled" : "Disabled");
-    }
-    if (auto* chain = ikSystem.getChain("RightArm")) {
-        ImGui::BulletText("Right Arm: %s", chain->enabled ? "Enabled" : "Disabled");
-    }
-    if (auto* chain = ikSystem.getChain("LeftLeg")) {
-        ImGui::BulletText("Left Leg: %s", chain->enabled ? "Enabled" : "Disabled");
-    }
-    if (auto* chain = ikSystem.getChain("RightLeg")) {
-        ImGui::BulletText("Right Leg: %s", chain->enabled ? "Enabled" : "Disabled");
-    }
-}
-
-void GuiSystem::renderSkeletonOverlay(Renderer& renderer, const Camera& camera) {
-    auto& sceneBuilder = renderer.getSceneBuilder();
-    if (!sceneBuilder.hasCharacter()) {
-        return;
-    }
-
-    auto& character = sceneBuilder.getAnimatedCharacter();
-
-    // Get the character's world transform from the scene object
-    const auto& sceneObjects = sceneBuilder.getRenderables();
-    size_t playerIndex = sceneBuilder.getPlayerObjectIndex();
-    if (playerIndex >= sceneObjects.size()) {
-        return;
-    }
-
-    glm::mat4 worldTransform = sceneObjects[playerIndex].transform;
-
-    // Get viewport size
-    float width = static_cast<float>(renderer.getWidth());
-    float height = static_cast<float>(renderer.getHeight());
-
-    // Get view-projection matrix
-    glm::mat4 viewProj = camera.getProjectionMatrix() * camera.getViewMatrix();
-
-    // Helper to project world position to screen
-    auto worldToScreen = [&](const glm::vec3& worldPos) -> ImVec2 {
-        glm::vec4 clipPos = viewProj * glm::vec4(worldPos, 1.0f);
-        if (clipPos.w <= 0.0f) {
-            return ImVec2(-1000, -1000);  // Behind camera
-        }
-        glm::vec3 ndc = glm::vec3(clipPos) / clipPos.w;
-        float screenX = (ndc.x * 0.5f + 0.5f) * width;
-        // Vulkan projection already flips Y (proj[1][1] *= -1), so NDC Y is already
-        // in screen orientation (negative = up, positive = down). Just map to pixels.
-        float screenY = (ndc.y * 0.5f + 0.5f) * height;
-        return ImVec2(screenX, screenY);
-    };
-
-    // Get the background draw list (renders behind everything)
-    ImDrawList* drawList = ImGui::GetBackgroundDrawList();
-
-    // Draw skeleton if enabled
-    if (ikDebugSettings.showSkeleton) {
-        SkeletonDebugData skelData = character.getSkeletonDebugData(worldTransform);
-
-        // Draw bones as lines
-        for (const auto& bone : skelData.bones) {
-            if (bone.parentIndex < 0) continue;  // Skip root
-
-            ImVec2 startScreen = worldToScreen(bone.startPos);
-            ImVec2 endScreen = worldToScreen(bone.endPos);
-
-            // Skip if behind camera
-            if (startScreen.x < -500 || endScreen.x < -500) continue;
-
-            // Bone color - cyan for normal, yellow for end effectors
-            ImU32 boneColor = bone.isEndEffector ?
-                IM_COL32(255, 255, 100, 200) :
-                IM_COL32(100, 255, 255, 200);
-
-            drawList->AddLine(startScreen, endScreen, boneColor, 2.0f);
-        }
-
-        // Draw joint positions as circles
-        for (const auto& pos : skelData.jointPositions) {
-            ImVec2 screenPos = worldToScreen(pos);
-            if (screenPos.x < -500) continue;
-
-            drawList->AddCircleFilled(screenPos, 4.0f, IM_COL32(255, 100, 100, 255));
-        }
-    }
-
-    // Draw IK targets if enabled
-    if (ikDebugSettings.showIKTargets) {
-        IKDebugData ikData = character.getIKDebugData();
-
-        // Transform IK positions to world space and draw
-        for (const auto& chain : ikData.chains) {
-            if (!chain.active) continue;
-
-            // Chain positions are already in skeleton local space
-            glm::vec4 targetWorld = worldTransform * glm::vec4(chain.targetPos, 1.0f);
-            glm::vec4 poleWorld = worldTransform * glm::vec4(chain.polePos, 1.0f);
-            glm::vec4 endWorld = worldTransform * glm::vec4(chain.endPos, 1.0f);
-
-            ImVec2 targetScreen = worldToScreen(glm::vec3(targetWorld));
-            ImVec2 poleScreen = worldToScreen(glm::vec3(poleWorld));
-            ImVec2 endScreen = worldToScreen(glm::vec3(endWorld));
-
-            // Draw target as green cross
-            if (targetScreen.x > -500) {
-                drawList->AddLine(
-                    ImVec2(targetScreen.x - 8, targetScreen.y),
-                    ImVec2(targetScreen.x + 8, targetScreen.y),
-                    IM_COL32(100, 255, 100, 255), 2.0f);
-                drawList->AddLine(
-                    ImVec2(targetScreen.x, targetScreen.y - 8),
-                    ImVec2(targetScreen.x, targetScreen.y + 8),
-                    IM_COL32(100, 255, 100, 255), 2.0f);
-            }
-
-            // Draw pole vector as blue diamond
-            if (poleScreen.x > -500) {
-                drawList->AddQuadFilled(
-                    ImVec2(poleScreen.x, poleScreen.y - 6),
-                    ImVec2(poleScreen.x + 6, poleScreen.y),
-                    ImVec2(poleScreen.x, poleScreen.y + 6),
-                    ImVec2(poleScreen.x - 6, poleScreen.y),
-                    IM_COL32(100, 100, 255, 200));
-            }
-
-            // Draw line from end effector to target
-            if (targetScreen.x > -500 && endScreen.x > -500) {
-                drawList->AddLine(endScreen, targetScreen, IM_COL32(255, 255, 100, 150), 1.0f);
-            }
-        }
-
-        // Draw look-at targets
-        for (const auto& lookAt : ikData.lookAtTargets) {
-            if (!lookAt.active) continue;
-
-            glm::vec4 targetWorld = worldTransform * glm::vec4(lookAt.targetPos, 1.0f);
-            glm::vec4 headWorld = worldTransform * glm::vec4(lookAt.headPos, 1.0f);
-
-            ImVec2 targetScreen = worldToScreen(glm::vec3(targetWorld));
-            ImVec2 headScreen = worldToScreen(glm::vec3(headWorld));
-
-            // Draw target as magenta circle
-            if (targetScreen.x > -500) {
-                drawList->AddCircle(targetScreen, 10.0f, IM_COL32(255, 100, 255, 255), 12, 2.0f);
-            }
-
-            // Draw line from head to target
-            if (targetScreen.x > -500 && headScreen.x > -500) {
-                drawList->AddLine(headScreen, targetScreen, IM_COL32(255, 100, 255, 150), 1.0f);
-            }
-        }
-
-        // Draw foot placement targets
-        for (const auto& foot : ikData.footPlacements) {
-            if (!foot.active) continue;
-
-            glm::vec4 groundWorld = worldTransform * glm::vec4(foot.groundPos, 1.0f);
-            glm::vec4 footWorld = worldTransform * glm::vec4(foot.footPos, 1.0f);
-
-            ImVec2 groundScreen = worldToScreen(glm::vec3(groundWorld));
-            ImVec2 footScreen = worldToScreen(glm::vec3(footWorld));
-
-            // Draw ground target as orange square
-            if (groundScreen.x > -500) {
-                drawList->AddRectFilled(
-                    ImVec2(groundScreen.x - 5, groundScreen.y - 5),
-                    ImVec2(groundScreen.x + 5, groundScreen.y + 5),
-                    IM_COL32(255, 150, 50, 200));
-            }
-
-            // Draw line from foot to ground target
-            if (groundScreen.x > -500 && footScreen.x > -500) {
-                drawList->AddLine(footScreen, groundScreen, IM_COL32(255, 150, 50, 150), 1.0f);
-            }
-        }
-    }
-
-    // Draw cape colliders if enabled
-    if (playerSettings.showCapeColliders) {
-        auto& cape = sceneBuilder.getPlayerCape();
-        if (cape.isInitialized()) {
-            CapeDebugData capeData = cape.getDebugData();
-
-            // Draw sphere colliders (orange circles)
-            for (const auto& sphere : capeData.spheres) {
-                ImVec2 centerScreen = worldToScreen(sphere.center);
-                if (centerScreen.x < -500) continue;
-
-                // Approximate screen radius based on depth
-                glm::vec4 clipCenter = viewProj * glm::vec4(sphere.center, 1.0f);
-                glm::vec4 clipEdge = viewProj * glm::vec4(sphere.center + glm::vec3(sphere.radius, 0, 0), 1.0f);
-                float screenRadius = 20.0f;  // Default size
-                if (clipCenter.w > 0.1f && clipEdge.w > 0.1f) {
-                    float ndcRadius = std::abs((clipEdge.x / clipEdge.w) - (clipCenter.x / clipCenter.w));
-                    screenRadius = ndcRadius * 0.5f * width;
-                    screenRadius = glm::clamp(screenRadius, 5.0f, 100.0f);
-                }
-
-                drawList->AddCircle(centerScreen, screenRadius, IM_COL32(255, 150, 50, 200), 16, 2.0f);
-            }
-
-            // Draw capsule colliders (green lines with circles at ends)
-            for (const auto& capsule : capeData.capsules) {
-                ImVec2 p1Screen = worldToScreen(capsule.point1);
-                ImVec2 p2Screen = worldToScreen(capsule.point2);
-
-                if (p1Screen.x < -500 && p2Screen.x < -500) continue;
-
-                // Approximate screen radius for capsule ends
-                glm::vec4 clipP1 = viewProj * glm::vec4(capsule.point1, 1.0f);
-                float screenRadius = 15.0f;
-                if (clipP1.w > 0.1f) {
-                    glm::vec4 clipEdge = viewProj * glm::vec4(capsule.point1 + glm::vec3(capsule.radius, 0, 0), 1.0f);
-                    if (clipEdge.w > 0.1f) {
-                        float ndcRadius = std::abs((clipEdge.x / clipEdge.w) - (clipP1.x / clipP1.w));
-                        screenRadius = ndcRadius * 0.5f * width;
-                        screenRadius = glm::clamp(screenRadius, 4.0f, 60.0f);
-                    }
-                }
-
-                // Draw capsule axis line
-                if (p1Screen.x > -500 && p2Screen.x > -500) {
-                    drawList->AddLine(p1Screen, p2Screen, IM_COL32(100, 255, 100, 200), 3.0f);
-                }
-
-                // Draw circles at capsule ends
-                if (p1Screen.x > -500) {
-                    drawList->AddCircle(p1Screen, screenRadius, IM_COL32(100, 255, 100, 200), 12, 2.0f);
-                }
-                if (p2Screen.x > -500) {
-                    drawList->AddCircle(p2Screen, screenRadius, IM_COL32(100, 255, 100, 200), 12, 2.0f);
-                }
-            }
-
-            // Draw attachment points (cyan diamonds)
-            for (const auto& attachPos : capeData.attachmentPoints) {
-                ImVec2 screenPos = worldToScreen(attachPos);
-                if (screenPos.x < -500) continue;
-
-                drawList->AddQuadFilled(
-                    ImVec2(screenPos.x, screenPos.y - 8),
-                    ImVec2(screenPos.x + 8, screenPos.y),
-                    ImVec2(screenPos.x, screenPos.y + 8),
-                    ImVec2(screenPos.x - 8, screenPos.y),
-                    IM_COL32(100, 255, 255, 220));
-            }
-        }
-    }
-}
-
-void GuiSystem::renderPlayerSection(Renderer& renderer) {
-    ImGui::Spacing();
-
-    auto& sceneBuilder = renderer.getSceneBuilder();
-    if (!sceneBuilder.hasCharacter()) {
-        ImGui::TextDisabled("No animated character loaded");
-        return;
-    }
-
-    // Cape section
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.9f, 0.5f, 0.5f, 1.0f));
-    ImGui::Text("CAPE");
-    ImGui::PopStyleColor();
-
-    ImGui::Checkbox("Enable Cape", &playerSettings.capeEnabled);
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Toggle cape visibility and simulation");
-    }
-
-    ImGui::Checkbox("Show Cape Colliders", &playerSettings.showCapeColliders);
-    if (ImGui::IsItemHovered()) {
-        ImGui::SetTooltip("Visualize body colliders used for cape collision");
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // Cape info
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.7f, 0.7f, 1.0f));
-    ImGui::Text("CAPE INFO");
-    ImGui::PopStyleColor();
-
-    ImGui::BulletText("Cloth simulation: Verlet integration");
-    ImGui::BulletText("Body colliders: Spheres + Capsules");
-    ImGui::BulletText("Attachments: Shoulders + Upper back");
 }
