+++ src/GuiTerrainTab.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class Renderer;
+
+namespace GuiTerrainTab {
+    void render(Renderer& renderer);
+}
