+++ src/GuiTimeTab.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class Renderer;
+
+namespace GuiTimeTab {
+    void render(Renderer& renderer);
+}
