+++ src/GuiWaterTab.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class Renderer;
+
+namespace GuiWaterTab {
+    void render(Renderer& renderer);
+}
