+++ src/GuiWeatherTab.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class Renderer;
+
+namespace GuiWeatherTab {
+    void render(Renderer& renderer);
+}
